Correção tipos de dados

+++ CustomCountries.API/Models/Country.cs
@@ -3,9 +3,9 @@
     public class Country
     {
         public string NumericCode { get; set; }
-        public float Area { get; set; }
-        public float Population { get; set; }
-        public float PopulationDensity { get; set; }
+        public decimal Area { get; set; }
+        public decimal Population { get; set; }
+        public decimal PopulationDensity { get; set; }
         public string Capital { get; set; }
     }
 }
